Move getcurrentDateAndTime to common.
